netlify/functions/requirements.txt
openai
pdfminer.six
chardet